# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Sick-Fansubs HTTP API
  version: "1"
  description: >-
    Accepted Sick-Fansubs wire contracts. Product APIs use /api/v1; operational
    health endpoints are intentionally outside product API versioning.

    Cross-cutting: every endpoint whose chain carries the forced-change gate
    (the authenticated reads chain and the authenticated chain) additionally
    answers 403 with problem type /problems/auth/password-change-required for
    sessions whose account carries the forced-change flag. The /api/v1/auth/
    subtree never carries the gate, so its paths stay reachable while flagged —
    including PUT /api/v1/auth/password, the sign-out pair, the session
    bootstrap, POST /api/v1/auth/reset-password, and
    POST /api/v1/auth/resend-verification.

    Cross-cutting: every documented unsafe operation (POST, PUT, PATCH,
    DELETE) requires the trusted origin and, when a session is present, the
    session-bound X-CSRF-Token header matching the csrfToken returned by the
    session bootstrap — a missing, malformed, or mismatched token is a 403.
    Safe methods never require it, and an unauthenticated unsafe request
    (for example sign-in, register, forgot-password, or reset-password) is
    protected by the trusted origin alone. An expired session has the
    clearing sf_session cookie written before its handler runs, and the
    operations whose handler clears the cookie — sign-out, sign-out-all,
...
    their own response.

...
servers:
...
paths:
...
  /api/v1/users/me/favorites/projects:
    get:
      operationId: listProjectFavorites
      summary: Current user's favorite projects
      description: >-
        Keyset-paginated list of the authenticated user's published project
        favorites, newest favorite first. Only
        still-published projects appear. Cursor namespace fpv1.
      parameters:
        - name: limit
          in: query
          required: false
          description: >-
            Page size. Integers from 1 through 100; absent means 20. A
            non-integer or empty value is a 400; an integer outside the
            range is a 422 with violation field "limit". Duplicate or
            unknown query parameters are rejected.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: after
          in: query
          required: false
          description: >-
            Opaque versioned cursor from a previous pageInfo.endCursor.
            At most 256 characters; any value that does not decode to a
            valid cursor is a 400.
          schema:
            type: string
            maxLength: 256
      responses:
        "200":
          description: Favorite projects with continuation metadata.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/PasswordChangeRequired"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/me/favorites/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Opaque content identifier (RFC 3986 unreserved, 1-64 chars).
        schema:
          type: string
          pattern: "^[A-Za-z0-9._~-]{1,64}$"
    get:
      operationId: getProjectFavoriteStatus
      summary: Whether the current user favorited a project
      description: >-
        200 {favorited} for a published project; the same masked 404 as the
        project detail endpoint otherwise.
      responses:
        "200":
          description: Favorite status for published content.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/PasswordChangeRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: addProjectFavorite
      summary: Add a project favorite (idempotent)
      description: >-
        Adds the published project to the current user's favorites — 204 on
        success and on repeat (idempotent). Requires the trusted origin and
        a valid X-CSRF-Token. Non-published or unknown ids answer the
        masked 404.
      responses:
        "204":
          description: Favorite added (or already present).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: removeProjectFavorite
      summary: Remove a project favorite (idempotent)
      description: >-
        Removes the project from the current user's favorites — 204 always
        (idempotent). Requires the trusted origin and a valid X-CSRF-Token.
      responses:
        "204":
          description: Favorite removed (or absent).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  # ── Current user follows ──
...
  /api/v1/users/me/follows/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Opaque content identifier (RFC 3986 unreserved, 1-64 chars).
        schema:
          type: string
          pattern: "^[A-Za-z0-9._~-]{1,64}$"
    get:
      operationId: getProjectFollowStatus
      summary: Whether the current user follows a project
      description: >-
        200 {following} for a published project; the same masked 404 as the
        project detail endpoint otherwise.
      responses:
        "200":
          description: Follow status for published content.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/PasswordChangeRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: addProjectFollow
      summary: Follow a project (idempotent)
      description: >-
        Adds the published project to the current user's follows — 204 on
        success and on repeat (idempotent). Requires the trusted origin and
        a valid X-CSRF-Token. Non-published or unknown ids answer the
        masked 404.
      responses:
        "204":
          description: Follow added (or already present).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: removeProjectFollow
      summary: Stop following a project (idempotent)
      description: >-
        Removes the project from the current user's follows — 204 always
        (idempotent). Requires the trusted origin and a valid X-CSRF-Token.
      responses:
        "204":
          description: Follow removed (or absent).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  # ── Current user sessions ──
...
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List published projects (newest first)
      description: >-
        Public keyset-paginated list of published projects. Ordering is
        publishedAt milliseconds descending, with the opaque id ascending as
        the final tie-breaker. Only projects with status
        "published" and a non-null published timestamp are returned — drafts
        and archived projects never appear. limit defaults to 20 and is
        capped at 100; after is an opaque, versioned
        cursor obtained from pageInfo.endCursor of a previous page. Project
        cursors carry a distinct namespace prefix and never cross endpoints.
      parameters:
        - name: limit
          in: query
          required: false
          description: >-
            Page size. Integers from 1 through 100; absent means 20. A
            non-integer or empty value is a 400; an integer outside the
            range is a 422 with violation field "limit". Duplicate or
            unknown query parameters are rejected.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: after
          in: query
          required: false
          description: >-
            Opaque versioned cursor from a previous pageInfo.endCursor.
            At most 256 characters; any value that does not decode to a
            valid project cursor is a 400.
          schema:
            type: string
            maxLength: 256
      responses:
        "200":
          description: One page of published projects, newest first.
          headers:
            X-Request-ID:
              description: Server-generated request correlation identifier.
              schema:
                type: string
            Cache-Control:
              description: API responses default to no-store.
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createProject
      summary: Create a project (admin+); slug generated from the title
      description: >-
        Staff content creation. Requires an
        admin or super-admin session, the trusted origin, and a valid
        X-CSRF-Token. The slug is GENERATED server-side from the title
        (Greek→latin transliteration, lowercase latin letters, digits, and
        single hyphens, collision-suffixed -2, -3, …; an untransliterable
        title falls back to an id-derived slug) — the create body has NO
        slug field, and an unknown slug key is a 400 under the strict body
        reader. The status
        select offers draft|published (default published; archived is an
        edit-time state only). thumbnailPath must be a validated
        storage-relative media reference (media/images/<2hex>/<32hex>.<ext>
        with the file present and the subdirectory equal to the id's first
        two hex characters) — the value POST /api/v1/media returns. A
        published create stamps publishedAt with the server clock. The
        downloads array is REQUIRED (1-20 rows; each row needs a name and
        at least one of magnetUrl/torrentUrl; every link must parse as an
        http(s) URL with a host or a magnet URI — 422
        {downloads[i].magnetUrl, invalidFormat} etc.). Body hard-limit
        512 KiB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Project created (slug generated).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            Location:
              description: The created resource's staff detail address (live
                for every status — the public detail masks drafts).
              schema:
                type: string
                pattern: "^/api/v1/admin/projects/[A-Za-z0-9._~-]{1,64}$"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffProject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/admin/projects:
    get:
      operationId: listStaffProjects
      summary: List every project for staff (all statuses), filterable
      description: >-
        Staff list: all statuses with badges, ordered updatedAt
        milliseconds descending with the opaque id ascending. The
        draft-visibility gate applies: non-published rows are visible only
        to their creator or a strictly-higher role (same-role peers never
        see each other's drafts). Requires a moderator+ session. The cursor
        namespace (apv1.) carries updatedAt in its timestamp slot.
        The optional status and q filters narrow the
        page; they never change the ordering, so a cursor minted under one
        filter stays a valid position under another.
      parameters:
        - name: limit
          in: query
          required: false
          description: >-
            Page size. Integers from 1 through 100; absent means 20. A
            non-integer or empty value is a 400; an integer outside the
            range is a 422 with violation field "limit", code "outOfRange".
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: after
          in: query
          required: false
          description: >-
            Opaque versioned cursor from pageInfo.endCursor (namespace
            apv1.). At most 256 characters; anything that does not decode is
            a 400.
          schema:
            type: string
            maxLength: 256
        - name: status
          in: query
          required: false
          description: >-
            Exact status filter. One of draft, published,
            archived; absent or empty means every status. Any other value is
            a 422 with violation field "status", code "invalidValue". The
            schema admits the blank form explicitly, so a validating client
            can send what the server accepts.
          schema:
            anyOf:
              - type: string
                enum: [draft, published, archived]
              - type: string
                const: ""
        - name: q
          in: query
          required: false
          description: >-
            Title filter: the query
            is split at every character that is not a letter or a digit, each
            word is folded (case-insensitive, Greek accents and final sigma
            ignored — the same folding as the search collation), and every
            word must appear somewhere in the title. So "dr stone" finds
            "Dr. Stone", matching what the public search does with the same
            input. The value is trimmed; more than 100 characters (runes) is a
            422 with violation field "q", code "maxLength" (the bound is
            checked first), and a value within the bound that has no letters
            and no digits means no filter rather than an error.
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: One page of the staff list.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffProjectList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/projects/{id}:
    get:
      operationId: getStaffProject
      summary: Staff detail of one project (any status) with the edit ETag
      description: >-
        Staff detail: the full row — slug, status, revision, both
        timestamps, creator/updater, and the storage-relative thumbnailPath
        the edit form round-trips. Emits the revision-backed strong ETag
        the update/delete send back as If-Match. The draft-visibility gate
        applies: invisible rows answer the same masked 404 as the public
        reads.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The staff-shaped row.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              description: The revision-backed strong ETag for If-Match.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffProject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  # ── Public content: search ──
...
components:
  headers:
    RequestID:
      description: Server-generated request correlation identifier.
      schema:
        type: string
    NoStoreCacheControl:
      description: API responses default to no-store.
      schema:
        type: string
        const: no-store
...
  schemas:
...
    FavoriteList:
      type: object
      description: Keyset collection envelope.
      additionalProperties: false
      required:
        - items
        - pageInfo
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteItem"
        pageInfo:
          $ref: "#/components/schemas/PageInfo"
...
  responses:
    BadRequest:
      description: Malformed request (400) — problem type /problems/bad-request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidCredentials:
      description: Generic 401 — problem type /problems/auth/invalid-credentials.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    PasswordChangeRequired:
      description: >-
        403 — problem type /problems/auth/password-change-required: the
        session's account carries the forced-change flag and the endpoint sits
        on a chain with the forced-change gate. The /api/v1/auth/ subtree never
        carries the gate, so its paths (password change, sign-out, session
        bootstrap, reset, resend-verification) stay reachable while flagged.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Validation:
      description: 422 with field violations — problem type /problems/validation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: 500 generic unexpected failure — problem type /problems/internal-error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"